README: Add note about deleting Gemfile.lock
When `Gemfile.lock` was generated by one version of the referenced
docker image, using it with a different version may fail because it
doesn't include the correct versions of the gems referenced in the file.
An easy fix for this is to delete the file before running the docker
process as this regenerates a working version of the lockfile.